#5 put both evenHIstory openedTickets in their own files. For code maintainability
